move new code in place; remove marked
Note improved markdown checking required changes to source markdown
