+++ tools/doc/package.json
@@ -7,13 +7,14 @@
     "node": ">=6"
   },
   "dependencies": {
-    "marked": "^0.3.5",
     "rehype-raw": "^2.0.0",
     "rehype-stringify": "^3.0.0",
+    "remark-html": "^7.0.0",
     "remark-parse": "^5.0.0",
     "remark-rehype": "^3.0.0",
     "unified": "^7.0.0",
     "unist-util-find": "^1.0.1",
+    "unist-util-select": "^1.5.0",
     "unist-util-visit": "^1.3.1"
   },
   "devDependencies": {
